test_write_db.py
"""
Test module for the 'write_db' module.

The 'write_db' module provides functions for writing pandas DataFrame to various file formats.

Functions:
- write_db(df, fn, **kwargs): Writes a pandas DataFrame to a file.
- write(df, db_file, frmt, **kwargs): Helper function to write a DataFrame to a file.

Usage Example:
    import write_db
    import pandas as pd

    # Create a DataFrame
    df = pd.DataFrame({'A': [1, 2, 3], 'B': [4, 5, 6]})
    
    # Write DataFrame to a file
    write_db.write_db(df, 'data.csv')

    # Further file writing options are available using the provided functions.
"""
import unittest
import os
import pandas as pd
from write_db import write_db, write

class TestWriteDB(unittest.TestCase):
    """
    Test suite for the 'write_db' module.
    """

    def test_write_db_success(self):
        """
        Test the write_db function with a successful write operation.
        """
        # Create a test DataFrame
        test_df = pd.DataFrame({'A': [1, 2, 3], 'B': [4, 5, 6]})
        
        # Specify the file name for writing
        test_file = 'test.csv'
        
        # Call the write_db function
        result = write_db(test_df, test_file)
        
        # Assert that the write operation was successful
        self.assertTrue(result)
        
        # Clean up temporary test file
        os.remove(test_file)

    def test_write_db_invalid_df(self):
        """
        Test the write_db function with an invalid DataFrame.
        """
        # Pass an invalid DataFrame
        invalid_df = None
        
        # Specify the file name for writing
        test_file = 'invalid.csv'
        
        # Call the write_db function
        result = write_db(invalid_df, test_file)
        
        # Assert that the write operation failed
        self.assertFalse(result)

    def test_write_db_empty_df(self):
        """
        Test the write_db function with an empty DataFrame.
        """
        # Create an empty DataFrame
        empty_df = pd.DataFrame()
        
        # Specify the file name for writing
        test_file = 'empty.csv'
        
        # Call the write_db function
        result = write_db(empty_df, test_file)
        
        # Assert that the write operation failed
        self.assertFalse(result)

    def testwrite(self):
        """
        Test the write helper function.
        """
        # Create a test DataFrame
        test_df = pd.DataFrame({'A': [1, 2, 3], 'B': [4, 5, 6]})
        
        # Specify the file name for writing
        test_file = 'test.csv'
        
        # Specify the format
        frmt = '.csv'
        
        # Call the write function
        write(test_df, test_file, frmt)
        
        # Assert that the file is created
        self.assertTrue(os.path.exists(test_file))
        
        # Clean up temporary test file
        os.remove(test_file)


if __name__ == '__main__':
    unittest.main()
